requirements.txt
Flask==2.0.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.27.0
Flask-JWT==0.3.2
pytest==7.0.1
Flask-Classful==0.14.2
